==> R/00_all.qmd <==
---
title: "Do It"
author: 
  - Rune Daucke (runda, s165493)
  - David Faurdal (dalofa, s144523)
  - Luisa Weisch (s233028)
date: 12/02-2024
format:
  html:
    embed-resources: true
editor: visual
execute:
  output-dir: "../results"
editor_options: 
  chunk_output_type: console
---

```{r}
#| label: loading_some_libs
# Loading library
library(quarto)
library(knitr)
library(rmarkdown)
```

{{< include 01_load.qmd >}}

{{< include 02_clean.qmd >}}

{{< include 03_augment.qmd >}}

{{< include 04_describe.qmd >}}

{{< include 05_PCA.qmd >}}

{{< include 06_diff_gene_expression.qmd >}}

{{< include 07_expression_of_neurodeg.qmd >}}
