==> R/04_describe.qmd <==
---
title: "Data Description"
author: 
  - Rune Daucke (runda, s165493)
  - David Faurdal (dalofa, s144523)
  - Luisa Weisch (s233028)
date: 12/02-2024
format:
  html:
    embed-resources: true
editor: visual
editor_options: 
  chunk_output_type: console
---

### Data Description

The data used in this project is comprised of Affymetrix microarray gene expression measurements collected from the blood of 20 healthy individuals and 40 patients diagnosed with sporadic Parkinsons Disease. Imporantly, the data was collected before patients started drug treatment.

After cleaning, the data set is comprised of the expression value for 22215 distinct genes:

```{r}
#| label: distinct_genes
#| eval: TRUE
#| echo: TRUE
#| message: FALSE
#| warning: FALSE
library("tidyverse")
df <- read_tsv("../data/02_dat_clean.tsv")

df |>
  summarise(unique_genes = n_distinct(Gene))
```

The summary statistics of expression across genes are as follows:

```{r}
#| label: summary_statistics
#| eval: TRUE
#| echo: TRUE
#| message: FALSE
#| warning: FALSE
df |>
  summarize(
    Exp_Mean = mean(Expression), na.rm = TRUE,
    Exp_SD = sd(Expression), na.rm = TRUE,
    Exp_Min = min(Expression), na.rm = TRUE,
    Exp_Max = max(Expression), na.rm = TRUE
  )
```

If we plot the distribution of gene expression values as boxplots per sample we can see that they have similar expression distributions and thus do not need to be normalized.

```{r}
#| label: normalization_check_plot
#| eval: TRUE
#| echo: TRUE
#| message: FALSE
#| warning: FALSE
#| fig.width: 8
#| fig.height: 6

# Create the boxplot to evaluate need for normalization and data spread
ggplot(df, aes(x = Sample, y = Expression, fill = Condition)) +
  geom_boxplot() +
  theme_minimal() +
  labs(title = "Gene Expression by Samples", x = "Sample", y = "Expression") +
  theme(axis.text.x = element_text(angle = 90, hjust = 1))

ggsave("../results/02_normalization_check.png", width = 8, height = 6, dpi = 600)
```
